requirements.txt
flask
line-bot-sdk
yfinance
